requirements.txt
openai>=1.0.0
anthropic>=0.39.0
google-generativeai>=0.8.0
ddgs>=9.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
prompt-toolkit>=3.0.0
